refactor: Setup as proper package
